third_party/blink/web_tests/platform/mac-mac15-arm64/virtual/webnn-service-on-cpu/external/wpt/webnn/conformance_tests/qdq_subgraph.https.any.worker_cpu-expected.txt

This is a testharness.js-based test.
[FAIL] [optional] quantized convTranspose2d
  assert_less_than_equal: assert_array_approx_equals_ulp: test quantizeLinear dequantizeLinear dequantizeLinear dequantizeLinear convTranspose2d quantizeLinear dequantizeLinear float32 actual 0.03529411926865578 should be close enough to expected 0.03921568766236305 by ULP distance: expected a number less than or equal to 8n but got 1052688n
Harness: the test ran to completion.
